+++ p3_frontend/src/symbolic_expression.rs
@@ -21,7 +21,6 @@
 // THE SOFTWARE.
 
 use alloc::boxed::Box;
-use alloc::rc::Rc;
 use core::fmt::Debug;
 use core::iter::{Product, Sum};
 use core::ops::{Add, AddAssign, Mul, MulAssign, Neg, Sub, SubAssign};
