+++ p3_frontend/tests/fib_air.rs
@@ -25,22 +25,10 @@
 use std::borrow::Borrow;
 
 use p3_air::{Air, AirBuilder, BaseAir};
-use p3_frontend::AirBuilderWithPublicValues;
-// use p3_baby_bear::{BabyBear, DiffusionMatrixBabybear};
-// use p3_challenger::DuplexChallenger;
-// use p3_commit::ExtensionMmcs;
-// use p3_dft::Radix2DitParallel;
-// use p3_field::extension::BinomialExtensionField;
 use p3_field::{AbstractField, Field, PrimeField};
-// use p3_fri::{FriConfig, TwoAdicFriPcs};
+use p3_frontend::AirBuilderWithPublicValues;
 use p3_matrix::dense::RowMajorMatrix;
 use p3_matrix::{Matrix, MatrixRowSlices};
-// use p3_merkle_tree::FieldMerkleTreeMmcs;
-// use p3_poseidon2::Poseidon2;
-// use p3_symmetric::{PaddingFreeSponge, TruncatedPermutation};
-// use p3_uni_stark::{prove, verify, StarkConfig};
-// use p3_util::log2_ceil_usize;
-// use rand::thread_rng;
 
 /// For testing the public values feature
 
@@ -232,71 +220,3 @@ fn test_fib() {
     .expect("verify succeeds");
     println!("Verify: {:?}", start.elapsed());
 }
-
-// type Val = BabyBear;
-// type Perm = Poseidon2<Val, DiffusionMatrixBabybear, 16, 7>;
-// type MyHash = PaddingFreeSponge<Perm, 16, 8, 8>;
-// type MyCompress = TruncatedPermutation<Perm, 2, 8, 16>;
-// type ValMmcs =
-//     FieldMerkleTreeMmcs<<Val as Field>::Packing, <Val as Field>::Packing, MyHash, MyCompress, 8>;
-// type Challenge = BinomialExtensionField<Val, 4>;
-// type ChallengeMmcs = ExtensionMmcs<Val, Challenge, ValMmcs>;
-// type Challenger = DuplexChallenger<Val, Perm, 16>;
-// type Dft = Radix2DitParallel;
-// type Pcs = TwoAdicFriPcs<Val, Dft, ValMmcs, ChallengeMmcs>;
-// type MyConfig = StarkConfig<Pcs, Challenge, Challenger>;
-//
-// #[test]
-// fn test_public_value() {
-//     let perm = Perm::new_from_rng(8, 22, DiffusionMatrixBabybear, &mut thread_rng());
-//     let hash = MyHash::new(perm.clone());
-//     let compress = MyCompress::new(perm.clone());
-//     let val_mmcs = ValMmcs::new(hash, compress);
-//     let challenge_mmcs = ChallengeMmcs::new(val_mmcs.clone());
-//     let dft = Dft {};
-//     let trace = generate_trace_rows::<Val>(0, 1, 1 << 3);
-//     let fri_config = FriConfig {
-//         log_blowup: 2,
-//         num_queries: 28,
-//         proof_of_work_bits: 8,
-//         mmcs: challenge_mmcs,
-//     };
-//     let pcs = Pcs::new(log2_ceil_usize(trace.height()), dft, val_mmcs, fri_config);
-//     let config = MyConfig::new(pcs);
-//     let mut challenger = Challenger::new(perm.clone());
-//     let pis = vec![
-//         BabyBear::from_canonical_u64(0),
-//         BabyBear::from_canonical_u64(1),
-//         BabyBear::from_canonical_u64(21),
-//     ];
-//     let proof = prove(&config, &FibonacciAir {}, &mut challenger, trace, &pis);
-//     let mut challenger = Challenger::new(perm);
-//     verify(&config, &FibonacciAir {}, &mut challenger, &proof, &pis).expect("verification failed");
-// }
-//
-// #[test]
-// #[should_panic(expected = "assertion `left == right` failed: constraints had nonzero value")]
-// fn test_incorrect_public_value() {
-//     let perm = Perm::new_from_rng(8, 22, DiffusionMatrixBabybear, &mut thread_rng());
-//     let hash = MyHash::new(perm.clone());
-//     let compress = MyCompress::new(perm.clone());
-//     let val_mmcs = ValMmcs::new(hash, compress);
-//     let challenge_mmcs = ChallengeMmcs::new(val_mmcs.clone());
-//     let dft = Dft {};
-//     let fri_config = FriConfig {
-//         log_blowup: 2,
-//         num_queries: 28,
-//         proof_of_work_bits: 8,
-//         mmcs: challenge_mmcs,
-//     };
-//     let trace = generate_trace_rows::<Val>(0, 1, 1 << 3);
-//     let pcs = Pcs::new(log2_ceil_usize(trace.height()), dft, val_mmcs, fri_config);
-//     let config = MyConfig::new(pcs);
-//     let mut challenger = Challenger::new(perm.clone());
-//     let pis = vec![
-//         BabyBear::from_canonical_u64(0),
-//         BabyBear::from_canonical_u64(1),
-//         BabyBear::from_canonical_u64(123_123), // incorrect result
-//     ];
-//     prove(&config, &FibonacciAir {}, &mut challenger, trace, &pis);
-// }
